backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.8.0
python-dotenv==1.0.1
langchain==0.2.14
openai==1.52.0
reportlab==4.0.8
scikit-learn==1.3.2
joblib==1.4.2
numpy==1.26.4
